requirements.txt
Django
beautifulsoup4
django-q
slackclient
django-import-export
